Parse mdx files
