+++ package.json
@@ -199,6 +199,11 @@
     "lodash.isempty": "4.4.0",
     "lodash.isobject": "3.0.2",
     "lodash.keyby": "4.6.0",
+    "mdast-util-from-markdown": "1.2.0",
+    "mdast-util-frontmatter": "1.0.0",
+    "mdast-util-mdx": "2.0.0",
+    "micromark-extension-frontmatter": "1.0.0",
+    "micromark-extension-mdxjs": "1.0.0",
     "prettier": "2.8.1",
     "react": "18.2.0",
     "react-dnd": "14.0.4",
@@ -217,7 +222,9 @@
     "ts-node": "10.9.1",
     "ts-toolbelt": "9.6.0",
     "typescript": "4.9.3",
-    "webpack": "5.74.0"
+    "unist-util-find": "1.0.2",
+    "webpack": "5.74.0",
+    "yaml": "2.2.0"
   },
   "peerDependencies": {
     "react": "^17.0.0 || ^18.0.0",
